IWF-472: Move helpers folder
