+++ integ/workflow/wait_for_state_completion/routers.go
@@ -1,7 +1,7 @@
 package wait_for_state_completion
 
 import (
-	"github.com/indeedeng/iwf/helpers"
+	"github.com/indeedeng/iwf/integ/helpers"
 	"github.com/indeedeng/iwf/service/common/ptr"
 	"log"
 	"net/http"
