+++ integ/workflow/wf_state_api_timeout/routers.go
@@ -3,7 +3,7 @@ package wf_state_api_timeout
 import (
 	"github.com/gin-gonic/gin"
 	"github.com/indeedeng/iwf/gen/iwfidl"
-	"github.com/indeedeng/iwf/helpers"
+	"github.com/indeedeng/iwf/integ/helpers"
 	"github.com/indeedeng/iwf/integ/workflow/common"
 	"log"
 	"net/http"
