+++ integ/workflow/wf_state_options_data_attributes_loading/routers.go
@@ -3,7 +3,7 @@ package wf_state_options_data_attributes_loading
 import (
 	"github.com/gin-gonic/gin"
 	"github.com/indeedeng/iwf/gen/iwfidl"
-	"github.com/indeedeng/iwf/helpers"
+	"github.com/indeedeng/iwf/integ/helpers"
 	"github.com/indeedeng/iwf/integ/workflow/common"
 	"github.com/indeedeng/iwf/service"
 	"github.com/stretchr/testify/assert"
